// src/modules/work-orders/domain/schemas/dto/response/work-orders.get.response.ts
export interface WorkOrderListItem {
  workOrderId: string;
  orderCode: string;
  origin: string;
  workTypeId: number;
  workTypeName: string;
  priorityId: number;
  clientId: string;
  clientName: string;
  personType: string;
  status: string;
  creationDate: string;
  assignationDate: string | null;
  completionDate: string | null;
  description: string | null;
  location: string;
  cadastralKey: string | null;
  coordinates: string | null;
  createdUserId: string;
  assignedUserId: string | null;
  completedUserId: string | null;
  metadata: string | null;
  isDeleted: boolean;
}